+++ tests/Api/ApiTest.php
@@ -158,6 +158,8 @@ class ApiTest extends TestCase
             [Entities\Contact\PhoneNumber::class],
             [Entities\CustomerCredential::class],
             [Entities\File::class],
+            [Entities\CreditMemo::class],
+            [Entities\CreditMemoAllocation::class],
             [Entities\Invoice::class],
             [Entities\InvoiceItem::class],
             [Entities\PaymentMethods\PaymentCardMethod::class, null],